Made separate api calls for Search and notification methods
